Stores reused calcs in normalization, removes normalization function
normalization function moved to plot, because it's necessary for tsne plot. When normalizer=None, uses internal sped up normalization, otherwise reverts to user desired normalization
