+++ src/doubletdetection/__init__.py
@@ -1,2 +1,2 @@
-from .doubletdetection import BoostClassifier, normalize_counts, load_mtx, load_10x_h5
+from .doubletdetection import BoostClassifier, load_mtx, load_10x_h5
 from .plot import *
